Добавлена функция вывода элементов списка
